BMP file, and BST classification.
Amit Lazar, Daniel Gutkind and I built an app that help to send packs according to the Bar-Code of each pack. We got BMP file, and we change it to digit numbers code. We classified the numbers off each pack by a BST (Binary Search Tree). A particular type of contain data structure that stores "items" (such as numbers, names etc.).
  During the project ,we present a few applications in C language, a popular sorting way (BST), and  good mathematical thinking.